+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
  - New XML conversion framework for performance improvements
 ### Changed
  - XmlUtils is now using the improved XML conversion framework 
+ - Better implementation of DictionaryExtensions.ToInverse
 
 ## [5.0.0] - 2018-09-14
 ### Added